adds option to resample audio on read
if the sample rate of the input doesnt match
the target sample rate needed for the network,
we can resample it using torchaudio.transforms.

to support this, modify all the program usage to
include `input_sample_rate` and `target_sample_rate`
options and remove the `sr` option.  The new rate args
expect to be in frequency, so //1000 prior when passing
into the models to get the right conv filters