fixes duplicate reporting and validation on accum
after switching to using the global_step, grad
accum was not accounted for in reporting.  for
example, if grad accum was 4, the reporting and
validation checks would be run 4 times in a row,
as the global step hasnt incremented.

this fixes that by recording the last step seen
and not running if we already ran